Implementation of hashtree-based path revocation

+++ endhost/sciond.py
@@ -24,7 +24,7 @@ from itertools import product
 
 # SCION
 from infrastructure.scion_elem import SCIONElement
-from lib.crypto.hash_chain import HashChain
+from lib.crypto.hash_tree import ConnectedHashTree
 from lib.defines import (
     PATH_FLAG_SIBRA,
     PATH_SERVICE,
@@ -58,8 +58,6 @@ class SCIONDaemon(SCIONElement):
     """
     # Max time for a path lookup to succeed/fail.
     TIMEOUT = 5
-    # Number of tokens the PS checks when receiving a revocation.
-    N_TOKENS_CHECK = 20
     # Time a path segment is cached at a host (in seconds).
     SEGMENT_TTL = 300
     MAX_SEG_NO = 5  # TODO: replace by config variable.
@@ -233,40 +231,45 @@ class SCIONDaemon(SCIONElement):
 
     def handle_revocation(self, pkt):
         rev_info = pkt.get_payload()
-        logging.debug("Received revocation:\n%s", str(rev_info))
-        # Verify revocation.
-#         if not HashChain.verify(rev_info.proof, rev_info.rev_token):
-#             logging.info("Revocation verification failed.")
-#             return
-        # Go through all segment databases and remove affected segments.
-        deletions = self._remove_revoked_pcbs(self.up_segments,
-                                              rev_info.rev_token)
-        deletions += self._remove_revoked_pcbs(self.core_segments,
-                                               rev_info.rev_token)
-        deletions += self._remove_revoked_pcbs(self.down_segments,
-                                               rev_info.rev_token)
-        logging.debug("Removed %d segments due to revocation.", deletions)
+        logging.debug("Received revocation:\n%s", rev_info)
 
-    def _remove_revoked_pcbs(self, db, rev_token):
+        # Go through all segment databases and remove affected segments.
+        self._remove_revoked_pcbs(self.up_segments, rev_info)
+        self._remove_revoked_pcbs(self.core_segments, rev_info)
+        self._remove_revoked_pcbs(self.down_segments, rev_info)
+
+    def _remove_revoked_pcbs(self, db, rev_info):
         """
         Removes all segments from 'db' that contain an IF token for which
         rev_token is a preimage (within 20 calls).
 
         :param db: The PathSegmentDB.
         :type db: :class:`lib.path_db.PathSegmentDB`
-        :param rev_token: The revocation token.
-        :type rev_token: bytes
+        :param rev_info: The revocation info
+        :type rev_info: RevocationInfo
 
         :returns: The number of deletions.
         :rtype: int
         """
+
+        if not ConnectedHashTree.verify_epoch(rev_info.p.epoch):
+            logging.warning("Epochs did not match")
+            return
+
         to_remove = []
         for segment in db():
-            for iftoken in segment.get_all_iftokens():
-                if HashChain.verify(rev_token, iftoken, self.N_TOKENS_CHECK):
+            for asm in segment.pcb.iter_asms():
+                if self.verify_asm(asm, rev_info):
                     to_remove.append(segment.get_hops_hash())
-
         return db.delete_all(to_remove)
+
+    def verify_asm(self, asm, rev_info):
+        ingress_if_id = asm.pcbm(0).hof().ingress_if
+        egress_if_id = asm.pcbm(0).hof().egress_if
+        root = asm.p.root
+        root_verify = ConnectedHashTree.verify(rev_info, root)
+        return (rev_info.p.ifID == ingress_if_id or
+                rev_info.p.ifID == egress_if_id) and root_verify
 
     def get_paths(self, dst_ia, flags=()):
         """Return a list of paths."""
